refactor: 💡 trading view to control which chart to display
