+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.1.0+25.25",
+  "version": "0.1.0+25.26",
   "private": true,
   "scripts": {
     "dev": "next dev",
